Enable a resin-like sanitization mechanism.
This adds a global DOM sanitization function which is settable via the setSanitizeDOMValue call.

If set, this function is called whenever lit-html would write an attribute, property, or text to the dom, with all of the information necessary to determine if the write is safe.

This API is analogous to the Polymer function of the same name, and it pairs nicely with https://github.com/Polymer/polymer-resin